fix: Avoid loading fee twice due to refresh button re-rendering twice issue
